% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashdog.R
\name{ashpen_fun}
\alias{ashpen_fun}
\title{Penalty on pivec used when \code{model = "ash"} in \code{\link{flexdog}}.}
\usage{
ashpen_fun(lambda, pivec)
}
\arguments{
\item{lambda}{The penalty.}

\item{pivec}{The vector of mixing proportions for the component
discrete uniform distributions.}
}
\value{
A penalty on the ash mixing weights.
}
\description{
Penalty on pivec used when \code{model = "ash"} in \code{\link{flexdog}}.
}
\author{
David Gerard
}
