% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convolve}
\alias{convolve}
\title{Convolution between two discrete probability mass functions
with support on 0:K.}
\usage{
convolve(x, y)
}
\arguments{
\item{x}{The first probability vector. The ith element is the
probability of i - 1.}

\item{y}{The second probability vector. The ith element is the
probability of i - 1.}
}
\value{
A vector that is the convolution of \code{x} and
    \code{y}. The ith element is the probability of i - 1.
}
\description{
Convolution between two discrete probability mass functions
with support on 0:K.
}
\author{
David Gerard
}
