% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dbernbinom}
\alias{dbernbinom}
\title{Special case of betabinomial where the beta is bernoulli mu.}
\usage{
dbernbinom(x, size, mu, log)
}
\arguments{
\item{x}{The quantile.}

\item{size}{The total number of draws.}

\item{mu}{The mean of the beta.}

\item{log}{A logical. Should we return the log of the
density \code{TRUE} or not \code{FALSE}?}
}
\value{
The density of the Bernoulli-binomial.
}
\description{
Special case of betabinomial where the beta is bernoulli mu.
}
\author{
David Gerard
}
