% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dbetabinom}
\alias{dbetabinom}
\title{Density function for the betabinomial distribution parameterized
by the mean and the overdispersion parameter.}
\usage{
dbetabinom(x, size, mu, rho, log)
}
\arguments{
\item{x}{A vector of quantiles.}

\item{size}{A vector of sizes.}

\item{mu}{Either a scalar of the mean for each observation,
or a vector of means of each observation, and thus
the same length as \code{x} and \code{size}. This must
be between 0 and 1.}

\item{rho}{Either a scalar of the overdispersion parameter
for each observation, or a vector of overdispersion
parameters of each observation, and thus the same length as
\code{x} and \code{size}. This must be between 0 and 1.}

\item{log}{A logical vector either of length 1 or the same
length as \code{x} and \code{size}. This determines whether
to return the log density for all observations (in the case
that its length is 1) or for each observation (in the case that
its length is that of \code{x} and \code{size}).}
}
\description{
This is the density function. It handles the corner cases when
rho is 0 or 1 in intelligible ways.
}
\details{
Let \eqn{\mu} and \eqn{\rho} be the mean and overdispersion paramters.
Let \eqn{\alpha} and \eqn{\beta} be the usual shape parameters of
a beta distribution. Then we have the relation
\deqn{\mu = \alpha/(\alpha + \beta),}
and
\deqn{\rho = 1/(1 + \alpha + \beta).}
This necessarily means that
\deqn{\alpha = \mu (1 - \rho)/\rho,}
and
\deqn{\beta = (1 - \mu) (1 - \rho)/\rho.}
}
\author{
David Gerard
}
