% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dbetabinom_double}
\alias{dbetabinom_double}
\title{The density function of the beta-binomial distribution.}
\usage{
dbetabinom_double(x, size, mu, rho, log)
}
\arguments{
\item{x}{The quantile.}

\item{size}{The total number of draws.}

\item{mu}{The mean of the beta.}

\item{rho}{The overdispersion parameter of the beta.}

\item{log}{A logical. Should we return the log of the
density \code{TRUE} or not \code{FALSE}?}
}
\value{
The density of the beta-binomial.
}
\description{
The density function of the beta-binomial distribution.
}
\author{
David Gerard
}
\keyword{internal}
