% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dc_dtau}
\alias{dc_dtau}
\title{Derivative of \eqn{c = (1 - \tau) / \tau} with respect to \eqn{\tau}.}
\usage{
dc_dtau(tau)
}
\arguments{
\item{tau}{The overdispersion parameter.}
}
\value{
A double.
}
\description{
Derivative of \eqn{c = (1 - \tau) / \tau} with respect to \eqn{\tau}.
}
\seealso{
\code{\link{dlbeta_dc}}, \code{\link{dlbeta_dtau}}
}
\author{
David Gerard
}
\keyword{internal}
