% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{df_deps}
\alias{df_deps}
\title{Derivative of f with respect to eps.}
\usage{
df_deps(p, eps)
}
\arguments{
\item{p}{The allele dosage.}

\item{eps}{The sequencing error rate.}
}
\value{
A double.
}
\description{
Derivative of f with respect to eps.
}
\author{
David Gerard
}
