% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dlbeta_dc}
\alias{dlbeta_dc}
\title{Derivative of the log-beta density with
respect to c where \eqn{c = (1 - \tau)/\tau}
where \eqn{\tau} is the overdispersion parameter.}
\usage{
dlbeta_dc(x, n, xi, c)
}
\arguments{
\item{x}{The number of successes observed}

\item{n}{The total number of trials observed.}

\item{xi}{The mean of the beta-binomial.}

\item{c}{\eqn{(1 - \tau)/\tau} where \eqn{\tau}
is the overdispersion parameter.}
}
\value{
A double.
}
\description{
Derivative of the log-beta density with
respect to c where \eqn{c = (1 - \tau)/\tau}
where \eqn{\tau} is the overdispersion parameter.
}
\seealso{
\code{\link{dbetabinom_double}}, \code{\link{dlbeta_dtau}},
    \code{\link{dc_dtau}}.
}
\author{
David Gerard
}
\keyword{internal}
