% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dlbeta_deps}
\alias{dlbeta_deps}
\title{Derivative of the log-beta-binomial density with respect to the
sequencing error rate.}
\usage{
dlbeta_deps(x, n, p, eps, h, tau)
}
\arguments{
\item{x}{The number of successes.}

\item{n}{The number of trials.}

\item{p}{The allele dosage.}

\item{eps}{The sequencing error rate}

\item{h}{The bias parameter.}

\item{tau}{The overdispersion parameter.}
}
\value{
A double.
}
\description{
Derivative of the log-beta-binomial density with respect to the
sequencing error rate.
}
\author{
David Gerard
}
\keyword{internal}
