% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dlbeta_dh}
\alias{dlbeta_dh}
\title{Derivative of log-betabinomial density with respect to bias parameter.}
\usage{
dlbeta_dh(x, n, p, eps, h, tau)
}
\arguments{
\item{x}{The number of successes.}

\item{n}{The number of trials.}

\item{p}{The allele dosage.}

\item{eps}{The sequencing error rate}

\item{h}{The bias parameter.}

\item{tau}{The overdispersion parameter.}
}
\value{
A double.
}
\description{
Derivative of log-betabinomial density with respect to bias parameter.
}
\author{
David Gerard
}
\keyword{internal}
