% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dlbeta_dtau}
\alias{dlbeta_dtau}
\title{Derivative of the log-beta-binomial density with respect to the
overdispersion parameter.}
\usage{
dlbeta_dtau(x, n, xi, tau)
}
\arguments{
\item{x}{The number of successes.}

\item{n}{The number of trials.}

\item{xi}{The mean of the beta.}

\item{tau}{The overdispersion parameter.}
}
\description{
Derivative of the log-beta-binomial density with respect to the
overdispersion parameter.
}
\seealso{
\code{\link{dlbeta_dc}}, \code{\link{dc_dtau}},
    \code{\link{dbetabinom_double}}.
}
\author{
David Gerard
}
