% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dlbeta_dxi}
\alias{dlbeta_dxi}
\title{Derivative of the log-betabinomial density with respect to the
mean of the underlying beta.}
\usage{
dlbeta_dxi(x, n, xi, tau)
}
\arguments{
\item{x}{The number of successes.}

\item{n}{The number of trials.}

\item{xi}{The mean of the underlying beta.}

\item{tau}{The overdispersion parameter.}
}
\value{
A double.
}
\description{
Derivative of the log-betabinomial density with respect to the
mean of the underlying beta.
}
\author{
David Gerard
}
