% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{dosage_to_gt}
\alias{dosage_to_gt}
\title{Convert the dosage to the common (unphased) GT format in VCF files.}
\usage{
dosage_to_gt(x, ploidy)
}
\arguments{
\item{x}{The allele dosage (proportion of alternative alleles).}

\item{ploidy}{The ploidy of the species.}
}
\value{
A character in the form 1/1/0/0 where 1 represents the alternative
    allele and 0 represents the reference allele.
}
\description{
Convert the dosage to the common (unphased) GT format in VCF files.
}
\author{
David Gerard
}
