% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{doutdist}
\alias{doutdist}
\title{The outlier distribution we use. Right now it is just a
beta binomial with mean 1/2 and od 1/3 (so underlying beta
is just a uniform from 0 to 1).}
\usage{
doutdist(x, n, logp)
}
\arguments{
\item{x}{The number of reference counts.}

\item{n}{The total number of read-counts.}

\item{logp}{Return the log density \code{TRUE} or not \code{FALSE}?}
}
\value{
A double. The outlier density value.
}
\description{
The outlier distribution we use. Right now it is just a
beta binomial with mean 1/2 and od 1/3 (so underlying beta
is just a uniform from 0 to 1).
}
\author{
David Gerard
}
