% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dpen_deps}
\alias{dpen_deps}
\title{Derivative of \deqn{-log(\epsilon(1 - \epsilon)) - (logit(\epsilon) - \mu_{\epsilon})^2 / (2\sigma_{\epsilon}^2)}
with respect to \eqn{\epsilon}.}
\usage{
dpen_deps(eps, mu_eps, sigma2_eps)
}
\arguments{
\item{eps}{The current sequencing error rate.}

\item{mu_eps}{The mean of the logit of the sequencing error rate.}

\item{sigma2_eps}{The variance of the logit of the sequencing error rate.}
}
\value{
A double.
}
\description{
Derivative of \deqn{-log(\epsilon(1 - \epsilon)) - (logit(\epsilon) - \mu_{\epsilon})^2 / (2\sigma_{\epsilon}^2)}
with respect to \eqn{\epsilon}.
}
\seealso{
\code{\link{pen_seq_error}} which this is a derivative for.
}
\author{
David Gerard
}
\keyword{internal}
