% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dpen_dh}
\alias{dpen_dh}
\title{Derivative of \deqn{-log(h) - (log(h) - \mu_h)^2 / (2\sigma_h^2)} with respect
to \eqn{h}.}
\usage{
dpen_dh(h, mu_h, sigma2_h)
}
\arguments{
\item{h}{The current bias parameter.}

\item{mu_h}{The mean of the log-bias.}

\item{sigma2_h}{The variance of the log-bias.}
}
\value{
A double.
}
\description{
Derivative of \deqn{-log(h) - (log(h) - \mu_h)^2 / (2\sigma_h^2)} with respect
to \eqn{h}.
}
\seealso{
\code{\link{pen_bias}} which this is a derivative for.
}
\author{
David Gerard
}
