% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_red.R
\name{dr_pen}
\alias{dr_pen}
\title{Penalty used in \code{\link{update_dr}}.}
\usage{
dr_pen(pairweights, mixing_pen)
}
\arguments{
\item{pairweights}{The mixing proportions to penalize.}

\item{mixing_pen}{The corresponding penalties.}
}
\description{
A dirichlet prior on \code{pairweights}. Returns log density.
}
\seealso{
\code{\link{update_dr}}
}
\author{
David Gerard
}
