% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dxi_df}
\alias{dxi_df}
\title{Derivative of xi with respect to f.}
\usage{
dxi_df(h, f)
}
\arguments{
\item{h}{The bias parameter.}

\item{f}{The post-sequencing error rate adjusted probability of an A.}
}
\value{
A double.
}
\description{
Derivative of xi with respect to f.
}
\author{
David Gerard
}
