% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dxi_dh}
\alias{dxi_dh}
\title{Derivative of xi-function with respect to bias parameter.}
\usage{
dxi_dh(p, eps, h)
}
\arguments{
\item{p}{The dosage (between 0 and 1).}

\item{eps}{The sequencing error rate.}

\item{h}{The bias parameter.}
}
\value{
A double.
}
\description{
Derivative of xi-function with respect to bias parameter.
}
\author{
David Gerard
}
