% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{elbo}
\alias{elbo}
\title{The evidence lower bound}
\usage{
elbo(warray, lbeta_array, cor_inv, postmean, postvar, bias, seq, mean_bias,
  var_bias, mean_seq, var_seq, ploidy)
}
\description{
The evidence lower bound
}
