% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eta_double}
\alias{eta_double}
\title{Adjusts allele dosage \code{p} by the sequencing error rate \code{eps}.}
\usage{
eta_double(p, eps)
}
\arguments{
\item{p}{The allele dosage.}

\item{eps}{The sequencing error rate.}
}
\value{
The probability of a reference reed adjusted by the
    sequencing error rate.
}
\description{
Adjusts allele dosage \code{p} by the sequencing error rate \code{eps}.
}
\author{
David Gerard
}
