% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eta_fun}
\alias{eta_fun}
\title{Adjusts allele dosage \code{p} by the sequencing error rate \code{eps}.}
\usage{
eta_fun(p, eps)
}
\arguments{
\item{p}{The allele dosage.}

\item{eps}{The sequencing error rate.}
}
\description{
Adjusts allele dosage \code{p} by the sequencing error rate \code{eps}.
}
