% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expit}
\alias{expit}
\title{The expit (logistic) function.}
\usage{
expit(x)
}
\arguments{
\item{x}{A double.}
}
\value{
The expit (logistic) of \code{x}.
}
\description{
The expit (logistic) function.
}
\author{
David Gerard
}
\keyword{internal}
