% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidog.R
\name{export_vcf}
\alias{export_vcf}
\title{Save the output of \code{\link{multidog}} as a VCF file.}
\usage{
export_vcf(obj, filename)
}
\arguments{
\item{obj}{An object of class \code{\link{multidog}}.}

\item{filename}{A string, the path to save the vcf file.}
}
\description{
Most of the output of \code{\link{multidog}} can be converted
to a VCF.
}
\details{
This function uses the Bioconductor packages VariantAnnotation,
GenomicRanges, S4Vectors, and IRanges. You can install these using the
BiocManager package via:

\code{# install.packages("BiocManager")}

\code{BiocManager::install(c("VariantAnnotation", "GenomicRanges", "S4Vectors", "IRanges"))}
}
