% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{f1_obj}
\alias{f1_obj}
\title{Objective for mixture of known dist and uniform dist.}
\usage{
f1_obj(alpha, pvec, weight_vec)
}
\arguments{
\item{alpha}{The mixing weight.}

\item{pvec}{The known distribtuion (e.g. from assuming an
F1 population).}

\item{weight_vec}{A vector of weights.}
}
\value{
The objective when updating \code{pivec} when \code{model = "f1"}
    or \code{model = "s1"} in \code{\link{flexdog_full}}.
}
\description{
Objective for mixture of known dist and uniform dist.
}
\author{
David Gerard
}
