% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashdog.R
\name{flexdog}
\alias{flexdog}
\title{Flexible updog}
\usage{
flexdog(refvec, sizevec, ploidy, model = c("ash", "flex"), verbose = TRUE,
  mean_bias = 0, var_bias = 1, mean_seq = -4.7, var_seq = 1,
  seq = 0.005, bias = 1, od = 0.001, mode = NULL)
}
\arguments{
\item{refvec}{A vector of counts of reads with reference allele.}

\item{sizevec}{A vector of total counts.}

\item{ploidy}{The ploidy of the species.}

\item{model}{What form should the prior take? Should the genotype distribution be unimodal
(\code{"ash"}) or should be generically any categorical distribution (\code{"flex"})?}

\item{verbose}{Should we output more (\code{TRUE}) or less (\code{FALSE})?}

\item{mean_bias}{The prior mean of the log-bias.}

\item{var_bias}{The prior variance of the log-bias.}

\item{mean_seq}{The prior mean of the logit of the sequencing error rate.}

\item{var_seq}{The prior variance of the logit of the sequencing error rate.}

\item{seq}{The starting value of the sequencing error rate.}

\item{bias}{The starting value of the bias.}

\item{od}{The starting value of the overdispersion parameter.}

\item{mode}{The mode if \code{model = "ash"}. If not provided, \code{flexdog} will estimate the mode.}
}
\description{
Flexible updog
}
\author{
David Gerard
}
