% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidog.R
\name{format_multidog}
\alias{format_multidog}
\title{Return matricized elements from the output of \code{\link{multidog}}.}
\usage{
format_multidog(x, varname = "geno")
}
\arguments{
\item{x}{The output of \code{multidog}.}

\item{varname}{The variable whose values populate the cells. This should
be a single columne from \code{x$inddf}.}
}
\description{
This function will allow you to have genotype estimates, maximum posterior
probability, and other values in the form of a matrix.
}
\author{
David Gerard
}
