% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_red.R
\name{get_bivalent_probs_dr}
\alias{get_bivalent_probs_dr}
\title{Double reduction version of \code{\link{get_bivalent_probs}}.}
\usage{
get_bivalent_probs_dr(ploidy)
}
\arguments{
\item{ploidy}{The ploidy of the individual. Should be even
and greater than 0.}
}
\value{
A list. The same elements as in \code{\link{get_bivalent_probs}}
    but with the additional element \code{dr_contrib_mat}. This is a logical 
    matrix where element (i, j) is \code{TRUE} if configuration i should be
    included if the parental genotype is j + 1, and \code{FALSE} if it should
    not be included. Inclusion is based on the double reduction model.
}
\description{
Double reduction version of \code{\link{get_bivalent_probs}}.
}
\seealso{
\code{\link{get_bivalent_probs}}.
}
\author{
David Gerard
}
