% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashdog.R
\name{get_dimname}
\alias{get_dimname}
\title{Returns a vector character strings that are all of the possible
combinations of the reference allele and the non-reference allele.}
\usage{
get_dimname(ploidy)
}
\arguments{
\item{ploidy}{The ploidy of the species.}
}
\value{
For example, if \code{ploidy = 3} then this will return
    c("aaa", "Aaa", "AAa", "AAA")
}
\description{
Returns a vector character strings that are all of the possible
combinations of the reference allele and the non-reference allele.
}
\author{
David Gerard
}
