% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_inner_weights}
\alias{get_inner_weights}
\title{Compute inner weights for updating the mixing proportions when using ash model.}
\usage{
get_inner_weights(ploidy, mode)
}
\arguments{
\item{ploidy}{The ploidy of the species.}

\item{mode}{The mode if \code{model = "ash"}. If not provided,
\code{flexdog} will estimate the mode.}
}
\description{
The (i,k)th element is \eqn{1(k \in F(a, i)) / |F(a,i)|}.
}
\author{
David Gerard
}
