% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_probk_vec}
\alias{get_probk_vec}
\title{Obtain the genotype distribution given the distribution of discrete uniforms.}
\usage{
get_probk_vec(pivec, model, mode)
}
\arguments{
\item{pivec}{The mixing probability of the i'th discrete uniform distribution.}

\item{model}{What form should the prior take? Should the genotype
distribution be unimodal (\code{"ash"}), generically
any categorical distribution (\code{"flex"}), or binomial as a
result of assuming Hardy-Weinberg equilibrium (\code{"hw"})?}

\item{mode}{The mode if \code{model = "ash"}. If not provided,
\code{flexdog} will estimate the mode. This is the starting point of
the allele frequency if \code{model = "hw"}.}
}
\description{
Obtain the genotype distribution given the distribution of discrete uniforms.
}
\seealso{
\code{\link{flexdog}} where this is used.
}
\author{
David Gerard
}
