% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_wik_mat}
\alias{get_wik_mat}
\title{E-step in \code{\link{flexdog}}.}
\usage{
get_wik_mat(probk_vec, refvec, sizevec, ploidy, seq, bias, od)
}
\arguments{
\item{probk_vec}{The vector of current prior probabilities of each genotype.}

\item{refvec}{A vector of counts of reads of the reference allele.}

\item{sizevec}{A vector of total counts.}

\item{ploidy}{The ploidy of the species. Assumed to be the same for each
individual.}

\item{seq}{The starting value of the sequencing error rate.}

\item{bias}{The starting value of the bias.}

\item{od}{The starting value of the overdispersion parameter.}
}
\value{
A matrix of numerics. The rows index the individuals and the
    columns index the genotype. These weights are used in the EM algorithm
    (and is indeed the E-step) in \code{\link{flexdog_full}}.
}
\description{
E-step in \code{\link{flexdog}}.
}
\seealso{
\code{\link{flexdog}} for the full EM algorithm.
}
\author{
David Gerard
}
\keyword{internal}
