% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_wik_mat_out}
\alias{get_wik_mat_out}
\title{E-step in \code{\link{flexdog}} where we now allow an outlier distribution.}
\usage{
get_wik_mat_out(probk_vec, out_prop, refvec, sizevec, ploidy, seq, bias, od)
}
\arguments{
\item{probk_vec}{The vector of current prior probabilities of each genotype.}

\item{out_prop}{The probability of being an outlier.}

\item{refvec}{A vector of counts of reads of the reference allele.}

\item{sizevec}{A vector of total counts.}

\item{ploidy}{The ploidy of the species. Assumed to be the same for each
individual.}

\item{seq}{The starting value of the sequencing error rate.}

\item{bias}{The starting value of the bias.}

\item{od}{The starting value of the overdispersion parameter.}
}
\value{
Same as \code{\link{get_wik_mat}} but the last column is
    for the outlier class.
}
\description{
E-step in \code{\link{flexdog}} where we now allow an outlier distribution.
}
\seealso{
\code{\link{flexdog}} for the full EM algorithm.
    \code{\link{get_wik_mat}} for the equivalent function
    without outliers. \code{\link{doutdist}} for the outlier
    density function.
}
\author{
David Gerard
}
\keyword{internal}
