% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad_for_eps}
\alias{grad_for_eps}
\title{Gradient for \code{\link{obj_for_eps}}.}
\usage{
grad_for_eps(parvec, refvec, sizevec, ploidy, mean_bias, var_bias, mean_seq,
  var_seq, wmat)
}
\arguments{
\item{parvec}{A vector of length three. The first element is the sequencing
error rate, the second element is the allele bias, and the third element
is the overdispersion parameter.}

\item{refvec}{A vector. The ith element is the reference count for the ith individual in the SNP.}

\item{sizevec}{A vector. the ith element is the size count for the ith individual in the SNP/}

\item{ploidy}{The ploidy of the species.}

\item{mean_bias}{The prior mean of the log-bias.}

\item{var_bias}{The prior variance of the log-bias}

\item{mean_seq}{The prior mean of the logit sequencing error rate.}

\item{var_seq}{The prior variance of the logit sequencing error rate.}

\item{wmat}{The matrix of variational posterior probabilities for each dosage.
The rows index the individuals and the columns index the dosage levels.}
}
\description{
Gradient for \code{\link{obj_for_eps}}.
}
\author{
David Gerard
}
