% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad_for_weighted_lbb}
\alias{grad_for_weighted_lbb}
\title{Gradient for \code{\link{obj_for_weighted_lbb}}.}
\usage{
grad_for_weighted_lbb(parvec, ploidy, weight_vec)
}
\arguments{
\item{parvec}{A vector of length 2. The first term is the current mean of the
underlying beta. The second term is the current overdispersion parameter.}

\item{ploidy}{The ploidy of the species.}

\item{weight_vec}{A vector of length \code{ploidy + 1} that contains the weights
for each component beta-binomial.}
}
\value{
A vector of length 2. The first component is the gradient of the mean of the underlying
    beta. The second component is the gradient of the overdispersion parameter of the
    underlying beta.
}
\description{
Gradient for \code{\link{obj_for_weighted_lbb}}.
}
\author{
David Gerard
}
\keyword{internal}
