% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad_for_weighted_lnorm}
\alias{grad_for_weighted_lnorm}
\title{Gradient for \code{\link{obj_for_weighted_lnorm}}.}
\usage{
grad_for_weighted_lnorm(parvec, ploidy, weight_vec)
}
\arguments{
\item{parvec}{A vector of length 2. The first term is the current mean of the
underlying normal. The second term is the current standard deviation
(not variance) of the normal.}

\item{ploidy}{The ploidy of the species.}

\item{weight_vec}{A vector of length \code{ploidy + 1} that contains the weights
for each component beta-binomial.}
}
\value{
A vector of length 2. The first term is the derivative with respect to the mean,
    the second term is the derivative with respect to the standard deviation (not variance).
}
\description{
Gradient for \code{\link{obj_for_weighted_lnorm}}.
}
\author{
David Gerard
}
\keyword{internal}
