% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashdog.R
\name{is.flexdog}
\alias{is.flexdog}
\title{Tests if an argument is a \code{flexdog} object.}
\usage{
is.flexdog(x)
}
\arguments{
\item{x}{Anything.}
}
\value{
A logical. \code{TRUE} if \code{x} is a \code{flexdog} object, and \code{FALSE} otherwise.
}
\description{
Tests if an argument is a \code{flexdog} object.
}
\examples{
is.flexdog("anything")
# FALSE

}
\author{
David Gerard
}
