% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidog.R
\name{is.multidog}
\alias{is.multidog}
\title{Tests if an argument is a \code{multidog} object.}
\usage{
is.multidog(x)
}
\arguments{
\item{x}{Anything.}
}
\value{
A logical. \code{TRUE} if \code{x} is a \code{multidog} object, and \code{FALSE} otherwise.
}
\description{
Tests if an argument is a \code{multidog} object.
}
\examples{
is.multidog("anything")
# FALSE

}
\author{
David Gerard
}
