% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vem.R
\name{is.mupdog}
\alias{is.mupdog}
\title{Tests if its argument is a mupdog object.}
\usage{
is.mupdog(x)
}
\arguments{
\item{x}{Anything.}
}
\value{
A logical. \code{TRUE} if \code{x} is a mupdog object, and \code{FALSE} otherwise.
}
\description{
Tests if its argument is a mupdog object.
}
\examples{
is.mupdog("anything")
# FALSE

}
\author{
David Gerard
}
