% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl.R
\name{kl_opt}
\alias{kl_opt}
\title{Finds the Gaussian-Binomial distribution closest (in terms of Kullback-Leibler divergence)
to a provided distribution.}
\usage{
kl_opt(dist, ploidy)
}
\arguments{
\item{dist}{A vector of probabilities. The \eqn{i}th element is the probability of having
\eqn{i+1} copies of the reference allele.}

\item{ploidy}{The ploidy of the individual.}
}
\value{
A list containing the fitted distribution along with the 
    fitted parameters.
}
\description{
The Gaussian-Binomial distribution is defined here by the hierarchical model
\deqn{z ~ N(0, \sigma^2),}
\deqn{y = F^{-1}(\Phi(z)|K, \alpha),}
where \eqn{F^{-1}(.|K, \alpha)} is the quantile function of a Binomial distribution with size
\eqn{K} and success probability \eqn{\alpha}. \eqn{\Phi(.)} is the standard normal cumulative
distribution function. In our context, \eqn{K} is the ploidy of the species. This function tries to
find the closest Gaussian-Binomial distribution to a provided distribution, where "closest" means
smallest Kulback-Leibler divergence. We do this by optimizing over both \eqn{\alpha} and \eqn{\sigma^2}.
}
\seealso{
\code{\link{kl_prob}} and \code{\link{kl_wrapp}} for the objective functions.
}
\author{
David Gerard
}
