% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl.R
\name{kl_prob}
\alias{kl_prob}
\title{KL-divergence between a provided distribution and a Gaussian-Binomial distribution.}
\usage{
kl_prob(dist, ploidy, sigma2, alpha)
}
\arguments{
\item{dist}{A vector of probabilities. The \eqn{i}th element is the probability of having
\eqn{i+1} copies of the reference allele.}

\item{ploidy}{The ploidy of the individual.}

\item{sigma2}{The underlying variance of the Gaussian-Binomial distribution.}

\item{alpha}{The probability of success.}
}
\description{
KL-divergence between a provided distribution and a Gaussian-Binomial distribution.
}
\seealso{
\code{\link{kl_opt}} for optimizing \code{kl_prob} over both \code{sigma2} and \code{alpha}.

    \code{\link{kl_wrapp}} for a wrapper of \code{kl_prob}.
}
\author{
David Gerard
}
