% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl.R
\name{kl_wrapp}
\alias{kl_wrapp}
\title{A wrapper of \code{\link{kl_prob}}.}
\usage{
kl_wrapp(par, dist, ploidy)
}
\arguments{
\item{par}{A vector of length two. The first element is \code{sigma2} in \code{\link{kl_prob}},
the second element is \code{alpha} in \code{\link{kl_prob}}.}

\item{dist}{A vector of probabilities. The \eqn{i}th element is the probability of having
\eqn{i+1} copies of the reference allele.}

\item{ploidy}{The ploidy of the individual.}
}
\description{
A wrapper of \code{\link{kl_prob}}.
}
\seealso{
\code{\link{kl_prob}} and \code{\link{kl_opt}} for optimizing \code{kl_wrapp} over \code{par}.
}
\author{
David Gerard
}
