% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_sum_exp}
\alias{log_sum_exp}
\title{Log-sum-exponential trick.}
\usage{
log_sum_exp(x)
}
\arguments{
\item{x}{A vector to log-sum-exp.}
}
\value{
The log of the sum of the exponential
    of the elements in \code{x}.
}
\description{
Log-sum-exponential trick.
}
\author{
David Gerard
}
