% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_sum_exp_2}
\alias{log_sum_exp_2}
\title{Log-sum-exponential trick using just two doubles.}
\usage{
log_sum_exp_2(x, y)
}
\arguments{
\item{x}{A double.}

\item{y}{Another double.}
}
\value{
The log of the sum of the exponential of x and y.
}
\description{
Log-sum-exponential trick using just two doubles.
}
\author{
David Gerard
}
