% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logit}
\alias{logit}
\title{The logit function.}
\usage{
logit(x)
}
\arguments{
\item{x}{A double between 0 and 1.}
}
\value{
The logit of \code{x}.
}
\description{
The logit function.
}
\author{
David Gerard
}
\keyword{internal}
