% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mupdog.R
\docType{package}
\name{mupdog-package}
\alias{mupdog-package}
\title{\code{mupdog}: An extenstion of updog to allow for correlation in the
    genotypes between individuals using a Gaussian copula model.}
\description{
\code{mupdog}: An extenstion of updog to allow for correlation in the
    genotypes between individuals using a Gaussian copula model.
}
\section{\code{mupdog} Functions}{

\itemize{
  \item \code{\link{mupdog}}: The main function in the package that fits the mupdog method.
}
}

\author{
David Gerard
}
