% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mupout}
\alias{mupout}
\title{A mupdog fit of the \code{\link{uitdewilligen}} data.}
\format{An object of class \code{\link{mupdog}}.}
\source{
The raw data that this was fit to can be found in \code{\link{uitdewilligen}}.
}
\usage{
mupout
}
\description{
A mupdog fit of the \code{\link{uitdewilligen}} data.
}
\seealso{
\code{\link{uitdewilligen}}, \code{\link{plot.mupdog}}, \code{\link{summary.mupdog}}
}
\keyword{datasets}
