% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{obj_for_alpha}
\alias{obj_for_alpha}
\title{Objective function when updating alpha}
\usage{
obj_for_alpha(mu, sigma2, alpha, rho, log_bb_dense, ploidy)
}
\arguments{
\item{mu}{A vector. The ith element is individual i's variational posterior mean at the SNP.}

\item{sigma2}{A vector. The ith element is individual i's variational posterior variance at the SNP.}

\item{alpha}{The SNP's allele frequency.}

\item{rho}{A vector. The ith element is individuals i's inbreeding coefficient.}

\item{log_bb_dense}{A matrix of log-densities of the beta binomial. The rows index the individuals and the columns index the allele dosage.}

\item{ploidy}{The ploidy of the species.}
}
\value{
A double. The objective when updating \code{alpha} in
    \code{\link{mupdog}}.
}
\description{
Objective function when updating alpha
}
\author{
David Gerard
}
