% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{obj_for_mu_wrapper}
\alias{obj_for_mu_wrapper}
\title{Wrapper for \code{\link{obj_for_mu}}.}
\usage{
obj_for_mu_wrapper(muSigma2Alpha, rho, log_bb_dense, ploidy, cor_inv)
}
\arguments{
\item{muSigma2Alpha}{A vector where the first \code{nsnps} observations are mu,
the next \code{nsnps} observations are sigma2, and the last observation is \code{alpha}.}

\item{rho}{A vector. The ith element is individuals i's inbreeding coefficient.}

\item{log_bb_dense}{A matrix of log-densities of the beta binomial. The rows index the individuals and the columns index the allele dosage.}

\item{ploidy}{The ploidy of the species.}

\item{cor_inv}{The inverse of the correlation matrix.}
}
\description{
Wrapper for \code{\link{obj_for_mu}}.
}
\author{
David Gerard
}
