% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{obj_for_weighted_lbb}
\alias{obj_for_weighted_lbb}
\title{Objective function for updating the beta-binomial genotype distribution when
\code{model = "bb"} in \code{\link{flex_update_pivec}}.}
\usage{
obj_for_weighted_lbb(parvec, ploidy, weight_vec)
}
\arguments{
\item{parvec}{A vector of length 2. The first term is the current mean of the
underlying beta. The second term is the current overdispersion parameter.}

\item{ploidy}{The ploidy of the species.}

\item{weight_vec}{A vector of length \code{ploidy + 1} that contains the weights
for each component beta-binomial.}
}
\value{
A double. The objective when updating the beta-binomial genotype
    distribution in \code{\link{mupdog}}.
}
\description{
Objective function for updating the beta-binomial genotype distribution when
\code{model = "bb"} in \code{\link{flex_update_pivec}}.
}
\author{
David Gerard
}
