% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{obj_for_weighted_lnorm}
\alias{obj_for_weighted_lnorm}
\title{Objective funtion for updating discrete normal genotype distribution
when \code{model = "normal"} in \code{\link{flex_update_pivec}}.}
\usage{
obj_for_weighted_lnorm(parvec, ploidy, weight_vec)
}
\arguments{
\item{parvec}{A vector of length 2. The first term is the current mean of the
underlying normal. The second term is the current standard deviation
(not variance) of the normal.}

\item{ploidy}{The ploidy of the species.}

\item{weight_vec}{A vector of length \code{ploidy + 1} that contains the weights
for each component beta-binomial.}
}
\value{
A double. The objective when updating the normal
    genotype distribution in \code{\link{mupdog}}.
}
\description{
Objective funtion for updating discrete normal genotype distribution
when \code{model = "normal"} in \code{\link{flex_update_pivec}}.
}
\author{
David Gerard
}
