% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{oracle_mis}
\alias{oracle_mis}
\title{Calculate oracle misclassification error.}
\usage{
oracle_mis(n, ploidy, seq, bias, od, dist)
}
\arguments{
\item{n}{The read-depth.}

\item{ploidy}{The ploidy of the individual.}

\item{seq}{The sequencing error rate.}

\item{bias}{The allele-bias}

\item{od}{The overdispersion parameter.}

\item{dist}{The distribution of the alleles.}
}
\description{
Given knowledge of the parameters, we calculate the expected misclassification error,
where the expectation is taken over both the data generation process and the allele-distribution.
This is an ideal level of misclassification error.
}
\details{
To come up with \code{dist}, you need some additional assumptions.
For example, if the population is in Hardy-Weinberg equilibrium and
the allele frequency is \code{alpha} then you could calculate
\code{dist} using the R code: \code{dbinom(x = 0:ploidy, size = ploidy, prob = alpha)}.
Alternatively, if you know the genotypes of the individual's two parents are, say,
\code{ref_count1} and \code{ref_count2}, then you could use the \code{\link[updog]{get_q_array}}
function from the updog package: \code{updog::get_q_array(ploidy)[ref_count1 + 1, ref_count2 + 1, ]}.
}
\author{
David Gerard
}
