% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pbetabinom_double}
\alias{pbetabinom_double}
\title{The distribution function of the betabinomial. This is generally
only advisable if q is relatively small.}
\usage{
pbetabinom_double(q, size, mu, rho, log_p)
}
\arguments{
\item{q}{A quantile.}

\item{size}{The total number of draws.}

\item{mu}{The mean of the beta.}

\item{rho}{The overdispersion parameter of the beta.}

\item{log_p}{A logical. Should return the log-probability
\code{TRUE} or not \code{FALSE}?}
}
\value{
The tail-probability of the beta-binomial.
}
\description{
The distribution function of the betabinomial. This is generally
only advisable if q is relatively small.
}
\author{
David Gerard
}
