% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pen_bias}
\alias{pen_bias}
\title{Penalty on bias parameter.}
\usage{
pen_bias(h, mu_h, sigma2_h)
}
\arguments{
\item{h}{The current value of bias parameter. Must be
greater than 0. A value of 1 means no bias.}

\item{mu_h}{The prior mean of the log-bias parameter.}

\item{sigma2_h}{The prior variance (not standard deviation)
of the log-bias parameter. Set to to \code{Inf} to return \code{0}.}
}
\value{
A double. The default penalty on the allelic bias parameter.
}
\description{
Penalty on bias parameter.
}
\author{
David Gerard
}
\keyword{internal}
