% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pen_seq_error}
\alias{pen_seq_error}
\title{Penalty on sequencing error rate.}
\usage{
pen_seq_error(eps, mu_eps, sigma2_eps)
}
\arguments{
\item{eps}{The current value of sequencing error rate.
Must be between 0 and 1.}

\item{mu_eps}{The prior mean of the logit sequencing error rate.}

\item{sigma2_eps}{The prior variance (not standard deviation)
of the logit sequencing error rate. Set this to \code{Inf} to
return \code{0}.}
}
\value{
A double. The default penalty on the sequencing error rate.
}
\description{
Penalty on sequencing error rate.
}
\author{
David Gerard
}
