% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pref.R
\name{pivec_from_segmats}
\alias{pivec_from_segmats}
\title{Function to get the segregation probabilities from the distributions
of each component and the weights of each component.}
\usage{
pivec_from_segmats(p1segmat, p2segmat, p1weight, p2weight)
}
\arguments{
\item{p1segmat}{The matrix of segregation probabilities for each
component for parent 1. The rows index the components and the
columns index the number of alleles to segregate.}

\item{p2segmat}{The matrix of segregation probabilities for each
component for parent 2. The rows index the components and the
columns index the number of alleles to segregate.}

\item{p1weight}{A vector of weights for each component (row) of
\code{p1segmat}.}

\item{p2weight}{A vector of weights for each component (row) of
\code{p2segmat}.}
}
\value{
A vector. The ith element is the probability of
    segregating i+1 total A alleles.
}
\description{
Function to get the segregation probabilities from the distributions
of each component and the weights of each component.
}
\seealso{
This is mostly used in \code{\link{update_pp}}.
}
\author{
David Gerard
}
