% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashdog.R
\name{plot.flexdog}
\alias{plot.flexdog}
\title{A wrapper for \code{\link[updog]{plot_geno}}. This will create a genotype plot.}
\usage{
\method{plot}{flexdog}(obj, ...)
}
\arguments{
\item{obj}{A \code{flexdog} object.}

\item{...}{Not used.}
}
\description{
A wrapper for \code{\link[updog]{plot_geno}}. This will create a genotype plot.
}
\details{
On a genotype plot, the x-axis contains the counts of the non-reference allele and the y-axis
contains the counts of the reference allele. The dashed lines are the expected counts (both reference and alternative)
given the sequencing error rate and the allele-bias. The plots are color-coded by the maximum-a-posterior genotypes.
Transparency is proportional to the maximum posterior probability for an
individual's genotype. Thus, we are less certain of the genotype of more transparent individuals.
}
\seealso{
\describe{
  \item{\code{\link[updog]{plot_geno}}}{The underlying plotting function.}
  \item{\code{\link{mupdog}}}{Creates a \code{mupdog} object.}
}
}
