% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{prob_to_phred}
\alias{prob_to_phred}
\title{Converts a proportion to a phred-scaled probability.}
\usage{
prob_to_phred(p)
}
\arguments{
\item{p}{A proportion. Between 0 and 1.}
}
\value{
\code{-10 * log(p, base = 10)}.
}
\description{
Converts a proportion to a phred-scaled probability.
}
\author{
David Gerard
}
