% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qbetabinom_double}
\alias{qbetabinom_double}
\title{The quantile function of the beta-binomial distribution parameterized
by mean and overdispersion parameter.}
\usage{
qbetabinom_double(p, size, mu, rho)
}
\arguments{
\item{p}{The lower tail probability.}

\item{size}{The total number of draws.}

\item{mu}{The mean of the beta.}

\item{rho}{The overdispersion parameter of the beta.}
}
\value{
The quantile of the beta-binomial.
}
\description{
The quantile function of the beta-binomial distribution parameterized
by mean and overdispersion parameter.
}
\author{
David Gerard
}
\keyword{internal}
