% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rbetabinom_int}
\alias{rbetabinom_int}
\title{One draw from the beta-binomial distribution parameterized
by mean and overdispersion parameter.}
\usage{
rbetabinom_int(size, mu, rho)
}
\arguments{
\item{size}{The total number of draws.}

\item{mu}{The mean of the beta.}

\item{rho}{The overdispersion parameter of the beta.}
}
\value{
A random sample from the beta-binomial.
}
\description{
One draw from the beta-binomial distribution parameterized
by mean and overdispersion parameter.
}
\author{
David Gerard
}
\keyword{internal}
