% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfs.R
\name{sfs_est}
\alias{sfs_est}
\title{Estimate SFS using algorithm of Li (2011)}
\usage{
sfs_est(glarray, tol = 1e-05, itermax = 1000)
}
\arguments{
\item{glarray}{An array of genotype log-likelihoods. Rows index SNPs,
columns index individuals, slices index genotypes}

\item{tol}{The stopping tolerance for the EM algorithm}

\item{itermax}{The maximum number of EM iterations.}
}
\value{
The estimated SFS.
}
\description{
We use the dynamic algorithm of Section 2.3.5 to obtain the site allele
frequency likelihood for all SNPs, then use the EM algorithm of Section
2.3.8 to estimate the site frequency spectrum.
}
\examples{
\donttest{
data("uitdewilligen", package = "updog")
mout <- multidog(
  refmat = uitdewilligen$refmat,
  sizemat = uitdewilligen$sizemat,
  ploidy = 4)

## estimate SFS using genotype likelihoods
glarray <- format_multidog(x = mout, varname = paste0("logL_", 0:4))
sfs_gl <- sfs_est(glarray)

## compare to SFS using posterior modes
genomat <- format_multidog(x = mout, varname = "geno")
sfs_mode <- unclass(
  prop.table(
    table(
      factor(
        rowSums(genomat),
        levels = 0:(ncol(genomat) * 4)
      )
     )
   )
 )

## pretty similar, some differences:
df <- data.frame(mode = sfs_mode, gl = sfs_gl)
round(df[df$mode > 1e-6 | df$gl > 1e-6, ], digits = 3)
}

}
\references{
\itemize{
  \item Li, H. (2011). A statistical framework for SNP calling, mutation discovery, association mapping and population genetical parameter estimation from sequencing data. Bioinformatics, 27(21), 2987-2993.
}
}
\author{
David Gerard
}
