% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfs.R
\name{sfs_pm}
\alias{sfs_pm}
\title{Bayesian estimate of the SFS}
\usage{
sfs_pm(postarray)
}
\arguments{
\item{postarray}{A 3-dimensional array of genotype posterior probabilities
(not log-probabilities). The rows index the SNPs, the columns
index the individuals, and the slices index the genotypes.}
}
\value{
The posterior mean SFS.
}
\description{
This function will calculate the posterior distribution of the
number of non-reference counts at each locus, then sum these
probabilities over loci to get the posterior mean of the SFS.
}
\examples{
\dontrun{
data("uitdewilligen", package = "updog")
mout <- multidog(
  refmat = uitdewilligen$refmat,
  sizemat = uitdewilligen$sizemat,
  ploidy = 4)

postarray <- format_multidog(x = mout, varname = paste0("Pr_", 0:4))

sfs <- sfs_pm(postarray)

}


}
\author{
David Gerard
}
