% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uitdewilligen}
\alias{uitdewilligen}
\title{Subset of individuals and SNPs from Uitdewilligen et al (2013).}
\format{
A list containing three objects. Two matrices and a numeric scalar:
\describe{
  \item{refmat}{A matrix of read counts containing the reference allele. The rows index the individuals and the columns index the SNPs.}
  \item{sizemat}{A matrix of the total number of read counts. The rows index the individuals and the columns index the SNPs.}
  \item{ploidy}{The ploidy level of the species (just 4).}
}
}
\source{
\url{https://doi.org/10.1371/journal.pone.0062355}
}
\usage{
uitdewilligen
}
\value{
A list. See the Format Section.
}
\description{
A list containing a matrix of reference counts, a matrix of total counts, and the ploidy level (4) of the species.
This is a subset of the data from Uitdewilligen et al (2013).
}
\references{
\itemize{
  \item{Uitdewilligen, J. G., Wolters, A. M. A., Bjorn, B., Borm, T. J., Visser, R. G., & van Eck, H. J. (2013). \href{https://doi.org/10.1371/journal.pone.0062355}{A next-generation sequencing method for genotyping-by-sequencing of highly heterozygous autotetraploid potato.} \emph{PLoS One}, 8(5), e62355.}
}
}
\keyword{datasets}
