% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{uni_obj_const}
\alias{uni_obj_const}
\title{Objective function optimized by \code{\link{uni_em_const}}.}
\usage{
uni_obj_const(pivec, alpha, weight_vec, lmat, lambda)
}
\arguments{
\item{pivec}{The current parameters.}

\item{alpha}{The mixing weight for the uniform component.
This should be small (say, less tahn 10^-3).}

\item{weight_vec}{A vector of weights. Each element of \code{weight_vec} corresponds
to a column of \code{lmat}.}

\item{lmat}{A matrix of inner weights. The columns are the "individuals" and the rows are the "classes."}

\item{lambda}{A vector of penalties on the pi's (corresponding to the rows
of \code{lmat}).
This can either be of length 1, in which case the same penalty is applied
to each of the pi's. Or it can be the same length of \code{pivec}, in
which case a different penalty is applied to each of the pi's. Larger
penalties generally increase the value of the pi's, not shrink them.}
}
\value{
The objective optimized by \code{\link{uni_em_const}} during
    that separate unimodal EM algorithm.
}
\description{
Objective function optimized by \code{\link{uni_em_const}}.
}
\author{
David Gerard
}
