% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashdog.R
\name{update_pp}
\alias{update_pp}
\title{Function to update the parameters in the preferential pairing model.}
\usage{
update_pp(weight_vec, model = c("f1pp", "s1pp"), control)
}
\arguments{
\item{weight_vec}{\code{colSums(wik_mat)} from \code{\link{flexdog}}.
This is the sum of current posterior probabilities of each individual
having genotype k.}

\item{model}{What model are we assuming? See the description in \code{\link{flexdog}} for details.}

\item{control}{A list of anything else needed to be passed.
E.g. if \code{model = "ash"},
then \code{inner_weights} needs to be passed through \code{control}
(see \code{\link{get_inner_weights}} for how to get this matrix).}
}
\value{
A list with the following elements:
\describe{
\item{}{}
}
}
\description{
Function to update the parameters in the preferential pairing model.
}
\author{
David Gerard
}
