% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vem.R
\name{update_R}
\alias{update_R}
\title{Update the underlying correlation matrix.}
\usage{
update_R(postmean, postvar)
}
\arguments{
\item{postmean}{The matrix of posterior means. The rows index the
individuals and the columns index the SNPs.}

\item{postvar}{The matrix of posterior variances. The rows index
the individuals and the columns index the SNPs.}
}
\description{
Update the underlying correlation matrix.
}
\author{
David Gerard
}
