% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updog.R
\docType{package}
\name{updog}
\alias{updog}
\alias{updog-package}
\title{\code{updog} Flexible Genotyping for Autopolyploids}
\description{
Implements empirical Bayes approaches to genotype
autopolyploids from next generation sequencing data while
accounting for allelic bias, overdispersion, and sequencing
error. The main function is \code{\link{flexdog}}, which
allows the specification
of many different genotype distributions. An experimental
function that takes into account varying leves of relatedness
is implemented in \code{\link{mupdog}}.
}
\section{\code{updog} Functions}{

\describe{
  \item{\code{\link{flexdog}}}{The main function that
      fits an empirical Bayes approach to genotype autopolyploids
      from next generation sequencing data.}
  \item{\code{\link{mupdog}}}{An experimental approach to genotype
      autopolyploids that accounts for varying levels of
      relatedness between the individuals in the sample.}
  \item{\code{\link{rgeno}}}{simulate the genotypes of a sample
      from one of the models allowed in \code{\link{flexdog}}.}
  \item{\code{\link{rflexdog}}}{Simulate from the
      \code{\link{flexdog}} model.}
  \item{\code{\link{plot.flexdog}}}{Plotting the output of
      \code{\link{flexdog}}.}
  \item{\code{\link{plot.mupdog}}}{Plotting the output of
      \code{\link{mupdog}}.}
  \item{\code{\link{summary.mupdog}}}{Providing some summaries
      of the output of \code{\link{mupdog}}.}
}
}

\section{\code{updog} Datasets}{

\describe{
  \item{\code{\link{snpdat}}}{A small example dataset for using
      \code{\link{flexdog}}.}
  \item{\code{\link{uitdewilligen}}}{A small example dataset
      for using \code{\link{mupdog}}.}
  \item{\code{\link{mupout}}}{The output from fitting
      \code{\link{mupdog}} to \code{\link{uitdewilligen}}.}
}
}

\author{
David Gerard
}
