% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{vcfdog}
\alias{vcfdog}
\title{Wrapper for \code{\link[VariantAnnotation]{readVcf}},
\code{\link[VariantAnnotation]{writeVcf}}, and
\code{\link{flexdog}}.}
\usage{
vcfdog(input, output, ploidy, nc = 1, alt_field = "AA", size_field = "DP",
  ...)
}
\arguments{
\item{input}{The input file. This file should be in VCF format.}

\item{output}{The ouptut file. This file will be in VCF format.}

\item{ploidy}{The ploidy of the species.}

\item{nc}{The number of cores to be use if parallelization
is desired. We use
the \code{\link[parallel]{parallel}} package
for parallel computing.}

\item{alt_field}{The name of the genotype field in \code{input}
that contains the alternative counts.}

\item{size_field}{The name of the genotype field in \code{input}
that contains the read-depth.}

\item{...}{Further arguements to pass to \code{\link{flexdog}}.}
}
\value{
The file \code{output} is created that
    contains the new INFO fields:
\describe{
\item{BIAS}{The allele (ascertainment) bias. The probability of ascertaining
    a reference read divided by the probability of ascertaining an alternative
    read.}
\item{SEQ}{The sequencing error rate.}
\item{OD}{The overdispersion parameter.}
\item{PM}{The posterior proportion of individuals misgenotyped.}
\item{PRIOR}{The estimated prior distribution. Comma separated.}
}
The \code{output} file also contains the new GENO fields:
\describe{
\item{GT}{The (unphased) gentotype. In the format 1/1/0/0 where 1 represents the
    alternative allele and 0 represents the reference allele.}
\item{DS}{The posterior mean dosage. Defined between 0 and \code{ploidy}.}
\item{GQ}{The posterior probability of misgenotyping.}
\item{GP}{The posterior probability of each genotype.}
}
}
\description{
Reads in a VCF file with \code{\link[VariantAnnotation]{readVcf}},
fits \code{\link{flexdog}} to each SNP in the VCF file, then
writes the output as a VCF file using
\code{\link[VariantAnnotation]{writeVcf}}.
}
\references{
Gerard, David, Luis Felipe Ventorim Ferrao,
    Antonio Augusto Franco Garcia, and Matthew Stephens. 2018.
    "Harnessing Empirical Bayes and Mendelian Segregation
    for Genotyping Autopolyploids from Messy Sequencing Data."
    \emph{bioRxiv}. Cold Spring Harbor Laboratory.
    doi:10.1101/281550.

    Valerie Obenchain, Michael Lawrence,
    Vincent Carey, Stephanie Gogarten, Paul Shannon,
    Martin Morgan; VariantAnnotation : a Bioconductor
    package for exploration and annotation of
    genetic variants , \emph{Bioinformatics}, Volume 30,
    Issue 14, 15 July 2014, Pages 2076-2078,
    \url{https://doi.org/10.1093/bioinformatics/btu168}.
}
\seealso{
\code{\link{flexdog}} for the underlying fitting function.
}
\author{
David Gerard
}
