% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xi_double}
\alias{xi_double}
\title{Adjusts allele dosage \code{p} by the sequencing error rate \code{eps} and the allele bias \code{h}.}
\usage{
xi_double(p, eps, h)
}
\arguments{
\item{p}{The allele dosage.}

\item{eps}{The sequencing error rate.}

\item{h}{The allele bias.}
}
\value{
The probability of a reference read adjusted by both the allele
    bias and the sequencing error rate.
}
\description{
Adjusts allele dosage \code{p} by the sequencing error rate \code{eps} and the allele bias \code{h}.
}
\author{
David Gerard
}
