% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xi_fun}
\alias{xi_fun}
\title{Adjusts allele dosage \code{p} by the sequencing error rate \code{eps} and the allele bias \code{h}.}
\usage{
xi_fun(p, eps, h)
}
\arguments{
\item{p}{A vector of allele dosages.}

\item{eps}{The sequencing error rate. Must be of length 1.}

\item{h}{The allele bias. Must be of length 1.}
}
\description{
Adjusts allele dosage \code{p} by the sequencing error rate \code{eps} and the allele bias \code{h}.
}
\author{
David Gerard
}
