% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dlbeta_dtau}
\alias{dlbeta_dtau}
\title{Derivative of the log-beta-binomial density with respect to the
overdispersion parameter.}
\usage{
dlbeta_dtau(x, n, p, eps, h, tau)
}
\arguments{
\item{x}{The number of successes.}

\item{n}{The number of trials.}

\item{p}{The allele dosage.}

\item{eps}{The sequencing error rate}

\item{h}{The bias parameter.}

\item{tau}{The overdispersion parameter.}
}
\value{
A double.
}
\description{
Derivative of the log-beta-binomial density with respect to the
overdispersion parameter.
}
\seealso{
\code{\link{dlbeta_dc}}, \code{\link{dc_dtau}},
    \code{\link{dbetabinom_double}}.
}
\author{
David Gerard
}
