% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eta_fun}
\alias{eta_fun}
\title{Adjusts allele dosage \code{p} by the sequencing error rate \code{eps}.}
\usage{
eta_fun(p, eps)
}
\arguments{
\item{p}{A vector of allele dosages.}

\item{eps}{The sequencing error rate. Must either be of length 1
or the same length as p.}
}
\value{
A vector of probabilities of a reference read adjusted
    by the sequencing error rate.
}
\description{
Adjusts allele dosage \code{p} by the sequencing error rate \code{eps}.
}
\author{
David Gerard
}
\keyword{internal}
