% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashdog.R
\name{flexdog}
\alias{flexdog}
\title{Flexible genotyping for autopolyploids from next-generation sequencing data.}
\usage{
flexdog(refvec, sizevec, ploidy, model = c("hw", "bb", "ash", "f1", "s1",
  "flex", "uniform"), verbose = TRUE, mean_bias = 0, var_bias = 0.7^2,
  mean_seq = -4.7, var_seq = 1, seq = 0.005, bias = 1, od = 0.001,
  update_bias = TRUE, update_seq = TRUE, update_od = TRUE, mode = NULL,
  use_cvxr = FALSE, itermax = 200, tol = 10^-4, fs1_alpha = 10^-3,
  ashpen = 0, p1ref = NULL, p1size = NULL, p2ref = NULL,
  p2size = NULL)
}
\arguments{
\item{refvec}{A vector of counts of reads with the reference allele.}

\item{sizevec}{A vector of total counts.}

\item{ploidy}{The ploidy of the species. Assumed to be the same for each
individual.}

\item{model}{What form should the prior take? Should the genotype
distribution be unimodal (\code{"ash"}), generically
any categorical distribution (\code{"flex"}), binomial as a
result of assuming Hardy-Weinberg equilibrium (\code{"hw"}),
an overdispersed binomial (\code{"bb"}),
a convolution of hypergeometics as a result that the population
consists of either an F1 cross (\code{"f1"}) or an S1
cross (\code{"s1"}), or fixed at a discrete uniform
(\code{"uniform"})?}

\item{verbose}{Should we output more (\code{TRUE}) or less
(\code{FALSE})?}

\item{mean_bias}{The prior mean of the log-bias.}

\item{var_bias}{The prior variance of the log-bias.}

\item{mean_seq}{The prior mean of the logit of the sequencing
error rate.}

\item{var_seq}{The prior variance of the logit of the sequencing
error rate.}

\item{seq}{The starting value of the sequencing error rate.}

\item{bias}{The starting value of the bias.}

\item{od}{The starting value of the overdispersion parameter.}

\item{update_bias}{A logical. Should we update \code{bias}
(\code{TRUE}), or not (\code{FALSE})?}

\item{update_seq}{A logical. Should we update \code{seq}
(\code{TRUE}), or not (\code{FALSE})?}

\item{update_od}{A logical. Should we update \code{od}
(\code{TRUE}), or not (\code{FALSE})?}

\item{mode}{The mode if \code{model = "ash"}. If not provided,
\code{flexdog} will estimate the mode. This is the starting point of
the allele frequency if \code{model = "hw"}. This should be
\code{NULL} for all other options of \code{model}.}

\item{use_cvxr}{A logical. If \code{model = "ash"}, then do you want
to use the EM algorithm
(\code{FALSE}) or a convex optimization program using
the package CVXR \code{TRUE}?
Only available if CVXR is installed.}

\item{itermax}{The maximum number of EM iterations to run for each mode
(if \code{model = "ash"}) or the total number of EM iterations to
run (if \code{model = "flex"}, \code{model = "hw"},
\code{model = "f1"}, \code{model = "s1"}, or \code{model = "uniform"}).}

\item{tol}{The tolerance stopping criterion. The EM algorithm will stop
if the difference in the log-likelihoods between two consecutive
iterations is less than \code{tol}.}

\item{fs1_alpha}{Either \code{"optim"} or the value at which to fix
the mixing proportion when \code{model = "f1"} or
\code{model = "s1"}. If \code{optim}, then we optimize over
the mixing proportion each iteration with possible values between
\code{10^-8} and \code{10^-3}. If you fix it, I would recommend some small
value such at \code{10^-3}.}

\item{ashpen}{The penalty to put on the unimodal prior.}

\item{p1ref}{The reference counts for the first parent if
\code{model = "f1"}, or for the only parent if \code{model = "s1"}.}

\item{p1size}{The total counts for the first parent if
\code{model = "f1"}, or for the only parent if \code{model = "s1"}.}

\item{p2ref}{The reference counts for the second parent if
\code{model = "f1"}.}

\item{p2size}{The total counts for the second parent if
\code{model = "f1"}.}
}
\value{
An object of class \code{flexdog}, which consists
    of a list with some or all of the following elements:
\describe{
  \item{\code{bias}}{The estimated bias parameter.}
  \item{\code{seq}}{The estimated sequencing error rate.}
  \item{\code{od}}{The estimated overdispersion parameter.}
  \item{\code{num_iter}}{The number of EM iterations ran. You should
      be wary if this equals \code{itermax}.}
  \item{\code{llike}}{The maximum marginal log-likelihood.}
  \item{\code{postmat}}{A matrix of posterior probabilities of each
      genotype for each individual. The rows index the individuals
      and the columns index the allele dosage.}
  \item{\code{gene_dist}}{The estimated genotype distribution. The
      \code{i}th element is the proportion of individuals with
      genotype \code{i-1}.}
  \item{\code{par}}{A list of the final estimates of the parameters
      of the genotype distribution. If \code{model = "hw"} then
      this will consist of \code{alpha}, the allele frequency.
      If \code{model = "f1"} or \code{model = "s1"} then this will
      consist of the parent genotype(s) and the mixing proportion
      with the discrete uniform (also called \code{alpha}).}
  \item{\code{geno}}{The posterior mode genotype.}
  \item{\code{maxpostprob}}{The maximum posterior probability.}
  \item{\code{postmean}}{The posterior mean genotype.}
  \item{\code{input$refvec}}{The value of \code{refvec} provided by
      the user.}
  \item{\code{input$sizevec}}{The value of \code{sizevec} provided
      by the user.}
  \item{\code{input$ploidy}}{The value of \code{ploidy} provided
      by the user.}
  \item{\code{input$model}}{The value of \code{model} provided by
      the user.}
  \item{\code{input$p1ref}}{The value of \code{p1ref} provided by the user.}
  \item{\code{input$p1size}}{The value of \code{p1size} provided by the user.}
  \item{\code{input$p2ref}}{The value of \code{p2ref} provided by the user.}
  \item{\code{input$p2size}}{The value of \code{p2size} provided by the user.}
  \item{\code{prop_mis}}{The posterior proportion of individuals
      misclassified.}
}
}
\description{
This function will genotype polyploid individuals from next generation
sequencing (NGS) data while assuming the genotype distribution is either
unimodal (\code{model = "ash"}), generically any categorical
distribution (\code{model = "flex"}), binomial as a result of assuming
the population is in Hardy-Weinberg equlibrium (\code{model = "hw"}),
an overdispersed binomial (\code{model = "bb"}),
results from the individuals being siblings from either an F1 cross
(\code{model = "f1"}) or an S1 cross (\code{model = "s1"}), or is a discrete
uniform distribution (\code{model = "uniform"}).
It does this while accounting for
allele bias, overdispersion, and sequencing error.
}
\details{
You might think a good default is \code{model = "uniform"} because it is
somehow an "uninformative prior." But it is very informative and tends to
work horribly in practice. I include it as an option only for completeness.

You might also think that allowing the genotype distribution to be anything
would work well (\code{model = "flex"}). But this tends to overfit the data
and get stuck in local modes during optimization.

A good default is either \code{model = "hw"}
if you have a small number of individuals (say, \eqn{<500}),
or \code{model = "ash"} if you have a lot of inidivuals (say, \eqn{>500}).
If the individuals are all siblings, then \code{model = "f1"} or
\code{model = "s1"} would work better. If the relatedness pattern between
individuals is more complicated, then I might recommend trying out
\code{\link{mupdog}}.

Prior to using \code{flexdog}, during the read-mapping step,
you could try to get rid of allelic bias by
using WASP (\url{https://doi.org/10.1101/011221}). If you are successful
in removing the allelic bias (because its only source was the read-mapping
step), then you could set \code{update_bias = FALSE}. You can visually
inspect SNPs for bias by using \code{\link[updog]{plot_geno}} from the
\code{updog} package.
}
\author{
David Gerard
}
