% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashdog.R
\name{flexdog}
\alias{flexdog}
\title{Flexible genotyping for autopolyploids from next-generation sequencing data.}
\usage{
flexdog(refvec, sizevec, ploidy, model = c("flex", "norm", "hw", "bb", "ash",
  "f1", "s1", "uniform"), p1ref = NULL, p1size = NULL, p2ref = NULL,
  p2size = NULL, bias_init = c(0.5, 0.7, 1, 1.4, 2), verbose = TRUE, ...)
}
\arguments{
\item{refvec}{A vector of counts of reads with the reference allele.}

\item{sizevec}{A vector of total counts.}

\item{ploidy}{The ploidy of the species. Assumed to be the same for each
individual.}

\item{model}{What form should the prior (genotype distribution) take?
See Details for possible values.}

\item{p1ref}{The reference counts for the first parent if
\code{model = "f1"}, or for the only parent if \code{model = "s1"}.}

\item{p1size}{The total counts for the first parent if
\code{model = "f1"}, or for the only parent if \code{model = "s1"}.}

\item{p2ref}{The reference counts for the second parent if
\code{model = "f1"}.}

\item{p2size}{The total counts for the second parent if
\code{model = "f1"}.}

\item{bias_init}{A vector of initial values for the bias parameter
over the multiple runs of \code{flexdog}.}

\item{verbose}{Should we output more (\code{TRUE}) or less
(\code{FALSE})?}

\item{...}{Additional parameters to pass to \code{\link{flexdog_full}}.}
}
\value{
An object of class \code{flexdog}, which consists
    of a list with some or all of the following elements:
\describe{
  \item{\code{bias}}{The estimated bias parameter.}
  \item{\code{seq}}{The estimated sequencing error rate.}
  \item{\code{od}}{The estimated overdispersion parameter.}
  \item{\code{num_iter}}{The number of EM iterations ran. You should
      be wary if this equals \code{itermax}.}
  \item{\code{llike}}{The maximum marginal log-likelihood.}
  \item{\code{postmat}}{A matrix of posterior probabilities of each
      genotype for each individual. The rows index the individuals
      and the columns index the allele dosage.}
  \item{\code{gene_dist}}{The estimated genotype distribution. The
      \code{i}th element is the proportion of individuals with
      genotype \code{i-1}.}
  \item{\code{par}}{A list of the final estimates of the parameters
      of the genotype distribution. If \code{model = "hw"} then
      this will consist of \code{alpha}, the allele frequency.
      If \code{model = "f1"} or \code{model = "s1"} then this will
      consist of the parent genotype(s) and the mixing proportion
      with the discrete uniform (also called \code{alpha}). If
      \code{model = "bb"} then this will consist of \code{alpha},
      the allele frequency, and \code{tau}, the overdispersion parameter.
      If \code{model = "norm"} then this will consist of \code{mu}, the
      normal mean, and \code{sigma}, the normal standard devation (not variance).}
  \item{\code{geno}}{The posterior mode genotype.}
  \item{\code{maxpostprob}}{The maximum posterior probability.}
  \item{\code{postmean}}{The posterior mean genotype.}
  \item{\code{input$refvec}}{The value of \code{refvec} provided by
      the user.}
  \item{\code{input$sizevec}}{The value of \code{sizevec} provided
      by the user.}
  \item{\code{input$ploidy}}{The value of \code{ploidy} provided
      by the user.}
  \item{\code{input$model}}{The value of \code{model} provided by
      the user.}
  \item{\code{input$p1ref}}{The value of \code{p1ref} provided by the user.}
  \item{\code{input$p1size}}{The value of \code{p1size} provided by the user.}
  \item{\code{input$p2ref}}{The value of \code{p2ref} provided by the user.}
  \item{\code{input$p2size}}{The value of \code{p2size} provided by the user.}
  \item{\code{prop_mis}}{The posterior proportion of individuals
      misclassified.}
  \item{\code{out_prop}}{The estimate proportion of points that are outliers.}
  \item{\code{prob_out}}{The ith element is the posterior probability
  that individual i is an outlier}
}
}
\description{
This function will genotype polyploid individuals from next generation
sequencing (NGS) data while assuming the genotype distribution is one of
several forms. It does this while accounting for allele bias, overdispersion,
and sequencing error.
}
\details{
Possible values of the genotype distribution (values of \code{model}) are:
\describe{
  \item{\code{"hw"}}{A binomial distribution that results from assuming that
      the population is in Hardy-Weinberg equilibrium (HWE). This actually does
      pretty well even when there are minor to moderate deviations from HWE.}
  \item{\code{"bb"}}{A beta-binomial distribution. This is a overdispersed
      version of "hw" and can be derived from a special case of the Balding-Nichols model.}
  \item{\code{"norm"}}{A distribution whose genotype frequencies are proportional
      to the density value of a normal with some mean and some standard deviation.
      Unlike the "bb" option, this will allow for distributions less dispersed than a binomial.}
  \item{\code{"ash"}}{Any unimodal prior. This will run \code{ploidy} EM algorithms
      with a different center during each optimization. It returns the center (and its fit)
      with the highest likelihood.}
  \item{\code{"f1"}}{This prior assumes the individuals are all full-siblings resulting
      from one generation of a bi-parental cross.}
  \item{\code{"s1"}}{This prior assumes the individuals are all full-siblings resulting
      from one generation of selfing. I.e. there is only one parent.}
  \item{\code{"flex"}}{Generically any categorical distribution. This works well if
      you have a lot of individuals, and so is the default.}
  \item{\code{"uniform"}}{A discrete uniform distribution. This should never be
      used in practice.}
}

You might think a good default is \code{model = "uniform"} because it is
somehow an "uninformative prior." But it is very informative and tends to
work horribly in practice. It will generally estimate the allele bias and sequencing
error rates to be unrealistic values so that the estimated genotypes are
approximately uniform. I include it as an option only for completeness.
Please don't use it.

Generally, good defaults are to use \code{model = "flex"} if you have
a lot of individuals (say, \eqn{> 10 * (ploidy + 1)}) and \code{model = "norm"} if you do not have
a lot of individuals (say, \eqn{< 10 * (ploidy + 1)}). This is if you use multiple initializations
of the bias as is the default in \code{\link{flexdog}}.

Prior to using \code{flexdog}, during the read-mapping step,
you could try to get rid of allelic bias by
using WASP (\url{https://doi.org/10.1101/011221}). If you are successful
in removing the allelic bias (because its only source was the read-mapping
step), then you could set \code{update_bias = FALSE} and \code{bias_init = 1}.
You can visually
inspect SNPs for bias by using \code{\link{plot_geno}}.
}
\seealso{
\code{\link{flexdog_full}} For additional parameter options.
}
\author{
David Gerard
}
