% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_red.R
\name{get_bivalent_probs_dr}
\alias{get_bivalent_probs_dr}
\title{Double reduction version of \code{\link{get_bivalent_probs}}.}
\usage{
get_bivalent_probs_dr(ploidy)
}
\arguments{
\item{ploidy}{The ploidy of the individual. Should be even
and greater than 0.}
}
\value{
A list. The same elements as in \code{\link{get_bivalent_probs}}, augmented to include
    more scenarios,
    but with the additional element \code{penvec}. This is a logical vector that
    is \code{TRUE} if the corresponding rows of \code{probmat} and \code{pmat} and
    elements of \code{lvec} would be
    included in the non-double-reduction-model and is \code{FALSE} otherwise.
}
\description{
Double reduction version of \code{\link{get_bivalent_probs}}.
}
\examples{
get_bivalent_probs(4)

}
\seealso{
\code{\link{get_bivalent_probs}}.
}
\author{
David Gerard
}
\keyword{internal}
