% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_probk_vec}
\alias{get_probk_vec}
\title{Obtain the genotype distribution given the distribution of discrete uniforms.}
\usage{
get_probk_vec(pivec, model, mode)
}
\arguments{
\item{pivec}{The mixing probability of the i'th discrete uniform distribution.}

\item{model}{What form should the prior (genotype distribution) take?
See Details for possible values.}

\item{mode}{The mode if \code{model = "ash"}. If not provided,
\code{flexdog} will estimate the mode. This is the starting point of
the allele frequency if \code{model = "hw"}. This should be
\code{NULL} for all other options of \code{model}.}
}
\value{
A vector of numerics. Element k is the probability of genotype k.
}
\description{
Obtain the genotype distribution given the distribution of discrete uniforms.
}
\seealso{
\code{\link{flexdog}} where this is used.
}
\author{
David Gerard
}
\keyword{internal}
