% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashdog.R
\name{initialize_pivec}
\alias{initialize_pivec}
\title{Initialize \code{pivec} for \code{\link{flexdog}} EM algorithm.}
\usage{
initialize_pivec(ploidy, mode, model = c("ash", "flex"))
}
\arguments{
\item{ploidy}{The ploidy of the species.}

\item{mode}{The mode if \code{model = "ash"}. If not provided,
\code{flexdog} will estimate the mode.}

\item{model}{What form should the prior take? Should the genotype
distribution be unimodal (\code{"ash"}) or should be generically
any categorical distribution (\code{"flex"})?}
}
\description{
The key idea here is choosing the pi's so that the two modes
have equal probability.
}
\seealso{
\code{\link{flexdog}} for where this is used.
}
\author{
David Gerard
}
