% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashdog.R
\name{initialize_pivec}
\alias{initialize_pivec}
\title{Initialize \code{pivec} for \code{\link{flexdog}} EM algorithm.}
\usage{
initialize_pivec(ploidy, mode, model = c("hw", "bb", "ash", "f1", "s1",
  "flex", "uniform"))
}
\arguments{
\item{ploidy}{The ploidy of the species. Assumed to be the same for each
individual.}

\item{mode}{The mode if \code{model = "ash"}. If not provided,
\code{flexdog} will estimate the mode. This is the starting point of
the allele frequency if \code{model = "hw"}. This should be
\code{NULL} for all other options of \code{model}.}

\item{model}{What form should the prior take? Should the genotype
distribution be unimodal (\code{"ash"}), generically
any categorical distribution (\code{"flex"}), binomial as a
result of assuming Hardy-Weinberg equilibrium (\code{"hw"}),
a convolution of hypergeometics as a result that the population
consists of either an F1 cross (\code{"f1"}) or an S1
cross (\code{"s1"}), or fixed at a discrete uniform
(\code{"uniform"})?}
}
\description{
The key idea here is choosing the pi's so that the two modes
have equal probability.
}
\seealso{
\code{\link{flexdog}} for where this is used.
}
\author{
David Gerard
}
