% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashdog.R
\name{initialize_pivec}
\alias{initialize_pivec}
\title{Initialize \code{pivec} for \code{\link{flexdog}} EM algorithm.}
\usage{
initialize_pivec(
  ploidy,
  mode,
  model = c("hw", "bb", "norm", "ash", "f1", "s1", "f1pp", "s1pp", "f1ppdr", "s1ppdr",
    "flex", "uniform", "custom")
)
}
\arguments{
\item{ploidy}{The ploidy of the species. Assumed to be the same for each
individual.}

\item{mode}{The mode if \code{model = "ash"}. If not provided,
\code{flexdog} will estimate the mode. This is the starting point of
the allele frequency if \code{model = "hw"}. This should be
\code{NULL} for all other options of \code{model}.}

\item{model}{What form should the prior (genotype distribution) take?
See Details for possible values.}
}
\value{
A vector of numerics. The initial value of \code{pivec}
    used in \code{\link{flexdog_full}}.
}
\description{
The key idea here is choosing the pi's so that the two modes
have equal probability.
}
\seealso{
\code{\link{flexdog}} for where this is used.
}
\author{
David Gerard
}
\keyword{internal}
