% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vem.R
\name{mupdog}
\alias{mupdog}
\title{Multivariate updog}
\usage{
mupdog(refmat, sizemat, ploidy, verbose = TRUE, mean_bias = 0,
  var_bias = 1, mean_seq = -4.7, var_seq = 1, seq = NULL, bias = NULL,
  od = NULL, allele_freq = NULL, inbreeding = NULL, cor_mat = NULL,
  postmean = NULL, postvar = NULL, update_cor = TRUE,
  update_inbreeding = TRUE, update_allele_freq = TRUE, num_clust = 1,
  control = list())
}
\arguments{
\item{refmat}{A matrix of reference counts.
The rows index the individuals and the columns index the SNPs.}

\item{sizemat}{A matrix of total counts.
The rows index the individuals and the columns index the SNPs.}

\item{ploidy}{The ploidy of the species. To estimate the ploidy,
re-run updog at various ploidy levels and choose the one with
the largest ELBO. This assumes that the ploidy is the same for
all individuals in the sample.}

\item{verbose}{Should we print a lot of output \code{TRUE}
or not \code{FALSE}?}

\item{mean_bias}{The prior mean of the log-bias. Defaults to 0
(no bias).}

\item{var_bias}{The prior variance on the log-bias. Defaults to 1.
This roughly corresponds to likely bias values between 0.14 and
7.4. This is a far wider interval than what we observe in
practice, thus making this prior rather uniformative. We usually
observe bias values somewhere between 0.5 and 2.}

\item{mean_seq}{The prior mean of the logit-sequencing-error-rate.
Defaults to -4.7. This corresponds to a sequencing error rate of
0.009.}

\item{var_seq}{The prior variance of the logit-sequencing-error-rate.
Defaults to 1. This corresponds to likely values of 0.001 and
0.06. This upper bound is larger than what we would expect given
the current state of next-gen-sequencings.}

\item{seq}{A vector of initial sequencing errors. Should be
the same length as the number of columns of \code{refmat}
(number of SNPs). Must be between 0 and 1.}

\item{bias}{A vector of initial bias paramters. Should be the
same length as the number of columns of \code{refmat}
(number of SNPs). Must be greater than 0.}

\item{od}{A vector of initial overdispersion parameters.
Should be the same length as the number of columns of
\code{refmat} (number of SNPs). Must be between 0 and 1.}

\item{allele_freq}{A vector of initial allele frequencies.
Should be the same length as the number of columns of
\code{refmat} (number of SNPs). Must be between 0 and 1.}

\item{inbreeding}{A vector of initial individual-specific
inbreeding coefficients.
Should be the same length as the number of rows of
\code{refmat} (number of individuals).}

\item{cor_mat}{Initial correlation matrix. Should have the same
number of columns/rows as the number of individuals.}

\item{postmean}{Initial variational posterior means. Should
have the same dimensions as \code{refmat}.}

\item{postvar}{Initial posterior variances. Should have the
same dimensions as \code{refmat}.}

\item{update_cor}{A logical. Should we update the underlying
correlation matrix \code{TRUE} or not \code{FALSE}. Will throw
an error if there are more individuals than SNPs and set to
\code{TRUE}.}

\item{update_inbreeding}{A logical. Should we update the
inbreeding coefficients \code{TRUE}
or not \code{FALSE}?}

\item{update_allele_freq}{A logical. Should we update the
allele frequencies \code{TRUE} or not \code{FALSE}?}

\item{num_clust}{The number of clusters to use if you want to
run the optimization steps in parallel. If \code{num_clust = 1},
then the optimization step will not be run in parallel.}

\item{control}{A list of control paramters (\code{itermax},
\code{obj_tol}).}
}
\description{
Multivariate updog
}
\examples{
\dontrun{
data(uitdewilligen)
mout <- mupdog(refmat = uitdewilligen$refmat,
               sizemat = uitdewilligen$sizemat,
               ploidy = uitdewilligen$ploidy,
               verbose = FALSE,
               control = list(obj_tol = 10^-5))
plot(mout, 4)
hist(mout$bias)
hist(mout$seq)
hist(mout$od)
hist(mout$inbreeding)
hist(mout$allele_freq)
}


}
\author{
David Gerard
}
