% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mupout}
\alias{mupout}
\title{A mupdog fit of the \code{\link{uitdewilligen}} data.}
\format{An object of class \code{\link{mupdog}}.}
\source{
The raw data that this was fit to can be found in \code{\link{uitdewilligen}}.
}
\usage{
mupout
}
\value{
See the Format Section.
}
\description{
A mupdog fit of the \code{\link{uitdewilligen}} data.
}
\seealso{
\describe{
\item{\code{\link{uitdewilligen}}}{The raw data.}
\item{\code{\link{plot.mupdog}}}{A method to plot a \code{\link{mupdog}} object.}
\item{\code{\link{summary.mupdog}}}{Calculate some summaries of a \code{\link{mupdog}} object.}
\item{\code{\link{mupdog}}}{Function used to create this \code{\link{mupdog}} object.}
}
}
\keyword{datasets}
