% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidog.R
\name{plot.multidog}
\alias{plot.multidog}
\title{Plot the output of \code{\link{multidog}}.}
\usage{
\method{plot}{multidog}(x, indices = seq(1, min(5, nrow(x$snpdf))), ...)
}
\arguments{
\item{x}{The output of \code{\link{multidog}}.}

\item{indices}{A vector of integers. The indices of the SNP's to plot.}

\item{...}{not used.}
}
\description{
Produce genotype plots from the output of \code{\link{multidog}}. You may
select which SNP's to plot.
}
\seealso{
\code{\link{plot_geno}}.
}
\author{
David Gerard
}
