% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mupdog.R
\name{plot.mupdog}
\alias{plot.mupdog}
\title{Draw a genotype plot from the output of \code{\link{mupdog}}.}
\usage{
\method{plot}{mupdog}(x, index, ...)
}
\arguments{
\item{x}{A mupdog object.}

\item{index}{The column number of the gene to plot.}

\item{...}{Not used.}
}
\description{
Draw a genotype plot from the output of \code{\link{mupdog}}.
}
\examples{
data(mupout)
plot(mupout, 4)

}
\author{
David Gerard
}
