% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xi_fun}
\alias{xi_fun}
\title{Adjusts allele dosage \code{p} by the sequencing error rate \code{eps} and the allele bias \code{h}.}
\usage{
xi_fun(p, eps, h)
}
\arguments{
\item{p}{A vector of allele dosages.}

\item{eps}{The sequencing error rate. Must either be of length 1
or the same length as p.}

\item{h}{The allele bias. Must either be of length 1 or the same length
as p.}
}
\value{
A vector of probabilities of the reference read adjusted
    by both the sequencing error rate and the allele bias.
}
\description{
Adjusts allele dosage \code{p} by the sequencing error rate \code{eps} and the allele bias \code{h}.
}
\author{
David Gerard
}
\keyword{internal}
